UPDATE NL 12/10/2024 19:00

+++ src/main/java/bnsshop/bnsshop/BrandServlet.java
@@ -9,7 +9,6 @@ import jakarta.servlet.http.HttpServletResponse;
 import models.Brand;
 import org.json.JSONObject;
 import utility.GestioneServlet;
-import utility.GestioneToken;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -99,6 +98,14 @@ public class BrandServlet extends HttpServlet{
 
     @Override
     public void doPut(HttpServletRequest request,HttpServletResponse response) throws ServletException,IOException{
+        String email = GestioneServlet.validaToken(request,response);
+        if (email.isEmpty()){
+            return;
+        }
+        if(!GestioneServlet.controllaRuolo(email)){
+            GestioneServlet.inviaRisposta(response,403,"\"Ruolo non corretto!\"",false);
+            return;
+        }
         int id= Integer.parseInt((String) request.getParameter("id"));
         BufferedReader reader=request.getReader();
         String row=reader.readLine();
@@ -127,6 +134,14 @@ public class BrandServlet extends HttpServlet{
 
     @Override
     public void doDelete(HttpServletRequest request,HttpServletResponse response) throws ServletException,IOException{
+        String email = GestioneServlet.validaToken(request,response);
+        if (email.isEmpty()){
+            return;
+        }
+        if(!GestioneServlet.controllaRuolo(email)){
+            GestioneServlet.inviaRisposta(response,403,"\"Ruolo non corretto!\"",false);
+            return;
+        }
         int id= Integer.parseInt((String) request.getParameter("id"));
         if (this.controller.deleteObject(id)){
             String message = "\"Product deleted Correctly.\"";
