+++ src/main/java/bnsshop/bnsshop/TagliaServlet.java
@@ -8,6 +8,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import models.Taglia;
 import org.json.JSONObject;
+import utility.GestioneServlet;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -74,6 +76,14 @@ public class TagliaServlet extends HttpServlet{
 
     @Override
     public void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException,IOException{
+        String email = GestioneServlet.validaToken(request,response);
+        if (email.isEmpty()){
+            return;
+        }
+        if(!GestioneServlet.controllaRuolo(email)){
+            GestioneServlet.inviaRisposta(response,403,"\"Ruolo non corretto!\"",false);
+            return;
+        }
         BufferedReader reader=request.getReader();
         String row=reader.readLine();
         List<String> rows = new ArrayList<>();
@@ -92,6 +102,14 @@ public class TagliaServlet extends HttpServlet{
 
     @Override
     public void doPut(HttpServletRequest request,HttpServletResponse response) throws ServletException,IOException{
+        String email = GestioneServlet.validaToken(request,response);
+        if (email.isEmpty()){
+            return;
+        }
+        if(!GestioneServlet.controllaRuolo(email)){
+            GestioneServlet.inviaRisposta(response,403,"\"Ruolo non corretto!\"",false);
+            return;
+        }
         int id= Integer.parseInt((String) request.getParameter("id"));
         BufferedReader reader=request.getReader();
         String row=reader.readLine();
@@ -133,6 +151,14 @@ public class TagliaServlet extends HttpServlet{
 
     @Override
     public void doDelete(HttpServletRequest request,HttpServletResponse response) throws ServletException,IOException{
+        String email = GestioneServlet.validaToken(request,response);
+        if (email.isEmpty()){
+            return;
+        }
+        if(!GestioneServlet.controllaRuolo(email)){
+            GestioneServlet.inviaRisposta(response,403,"\"Ruolo non corretto!\"",false);
+            return;
+        }
         int idTaglia= Integer.parseInt((String) request.getAttribute("idtaglia"));
         if (this.controller.deleteObject(idTaglia)){
             String message = "Product deleted Correctly.";
